Moved connect to index.
