+++ automation/helpers/scenes/index.js
@@ -1,1 +1,1 @@
-export { Home } from './Home';
+export Home from './Home';
