+++ source/comps/scenes/index.js
@@ -1,5 +1,5 @@
 export { Browse } from './Browse';
-export { Home } from './Home';
+export Home from './Home';
 export { Profile } from './Profile';
 export Submit from './Submit';
 export Inspect from './Inspect';
